fix error

+++ app/src/main/java/com/example/cicipinapp/models/ReviewModel.kt
@@ -3,7 +3,7 @@ package com.example.cicipinapp.models
 data class ReviewModel(
     val id: Int,
     val review: String,
-    val rating: Int,
+    val rating: Float,
     val description: String,
     val date: String,
     val RestaurantsID: Int
@@ -11,7 +11,7 @@ data class ReviewModel(
 
 data class CreateReviewRequest(
     val review: String,
-    val rating: Int,
+    val rating: Float,
     val description: String,
     val date: String,
     val RestaurantsID: Int
@@ -20,7 +20,7 @@ data class CreateReviewRequest(
 data class UpdateReviewRequest(
     val id: Int,
     val review: String,
-    val rating: Int,
+    val rating: Float,
     val description: String,
     val RestaurantsID: Int
 )
